requirements.txt
asgiref==3.6.0
astroid==2.14.2
dill==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.0.0
pylint==2.16.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
wrapt==1.15.0
